+++ priv/repo/migrations/20190711042024_copy_muted_to_muted_notifications.exs
@@ -3,22 +3,6 @@ defmodule Pleroma.Repo.Migrations.CopyMutedToMutedNotifications do
   alias Pleroma.User
 
   def change do
-    query =
-      User.Query.build(%{
-        local: true,
-        active: true,
-        order_by: :id
-      })
-
-    Pleroma.Repo.stream(query)
-    |> Enum.each(fn
-      %{info: %{mutes: mutes} = info} = user ->
-        info_cng =
-          Ecto.Changeset.cast(info, %{muted_notifications: mutes}, [:muted_notifications])
-
-        Ecto.Changeset.change(user)
-        |> Ecto.Changeset.put_embed(:info, info_cng)
-        |> Pleroma.Repo.update()
-    end)
+  execute("update users set info = jsonb_set(info, '{muted_notifications}', info->'mutes', true) where local = true")
   end
 end